backend/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
pyjwt==2.8.0
bcrypt==4.1.2
